+++ api/TeamLunch/Commands/NewFineRequest.cs
@@ -1,6 +1,4 @@
-using MauticNetClient;
-using MauticNetClient.Commands.Emails;
-using MauticNetClient.Models;
+
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using TeamLunch.Contracts;
@@ -25,9 +23,9 @@ public static class NewFineRequest
     {
         private readonly DataContext _db;
         private readonly INotificationService _notificationService;
-        private readonly EmailService _emailService;
+        private readonly IEmailService _emailService;
 
-        public Handler(DataContext db, INotificationService notificationService, EmailService emailService)
+        public Handler(DataContext db, INotificationService notificationService, IEmailService emailService)
         {
             _db = db;
             _notificationService = notificationService;