+++ web/e2e/authentication.spec.js
@@ -1,14 +1,13 @@
 const { test, expect } = require('@playwright/test');
 
-test('Authentication works with email and password', async ({ page }) => {
-  await page.goto('https://thankful-sand-0a1eb4203.1.azurestaticapps.net/');
-    
-  await page.locator('[placeholder="Email Address"]').click();
-  await page.locator('[placeholder="Email Address"]').fill(process.env.TEST_USER);
+test('test', async ({ page }) => {
+  await page.goto('https://doinfine.app/');
 
-  await page.locator('[placeholder="Password"]').click();
-  await page.locator('[placeholder="Password"]').fill(process.env.TEST_PASSWORD);
-  await page.locator('button:has-text("Sign in")').click();
+  await page.locator('input[type="email"]').click();
+  await page.locator('input[type="email"]').fill('billy@example.com');
+  await page.locator('input[type="email"]').press('Tab');
+  await page.locator('input[type="password"]').fill('#Billy22*');
+  await page.locator('text=SignIn').click();
 
-  await expect(page).toHaveURL('https://thankful-sand-0a1eb4203.1.azurestaticapps.net');
+  await expect(page).toHaveURL('https://doinfine.app/team');
 });