+++ web/src/components/molecules/LinkListItem/LinkListItem.js
@@ -1,18 +1,23 @@
-import { Divider, useTheme } from '@mui/material'
+import { Divider, ListItemButton, useTheme } from '@mui/material'
 import { ChevronRight } from '@mui/icons-material'
 import { Box, Chip, IconButton, ListItem, Typography } from '@mui/material'
 import React from 'react'
-import { Link } from 'react-router-dom'
+import { Link, useNavigate } from 'react-router-dom'
 
 export const LinkListItem = ({ label, link, handleLinkClick }) => {
     const theme = useTheme()
+    const navigate = useNavigate()
+
+    const handleClick = () => {
+        link ? navigate(link) : handleLinkClick()
+    }
 
     return (
         <>
-            <ListItem sx={{
+            <ListItemButton onClick={handleClick} sx={{
                 justifyContent: 'space-between',
                 alignItems: 'center',
-                padding: '0.5rem'
+                padding: '1rem 0.5rem'
             }}>
                 <Box sx={{
                     display: 'flex',
@@ -20,17 +25,9 @@ export const LinkListItem = ({ label, link, handleLinkClick }) => {
                 }}>
                     <Typography variant='p'>{label}</Typography>
                 </Box>
-                {handleLinkClick
-                    ? <IconButton onClick={handleLinkClick}>
-                        <ChevronRight />
-                    </IconButton>
-                    : <Link to={link}>
-                        <IconButton>
-                            <ChevronRight />
-                        </IconButton>
-                    </Link>
-                }
-            </ListItem>
+
+                <ChevronRight />
+            </ListItemButton>
             <Divider />
         </>
     )