+++ web/src/components/molecules/LinkListItem/LinkListItem.test.js
@@ -9,20 +9,8 @@ describe('LinkListItem', () => {
         renderWithRouter(<LinkListItem label={EXAMPLE_TITLE} link={EXAMPLE_LINK} />)
 
         const label = screen.getByText(EXAMPLE_TITLE)
-        const link = screen.getByRole('link')
-        const button = screen.getByRole('button')
 
         expect(label).toBeInTheDocument()
-        expect(button).toBeInTheDocument()
-        expect(link.attributes.getNamedItem('href').value).toBe(EXAMPLE_LINK)
-    })
-
-    it('renders with badge if count value is passed in', () => {
-        renderWithRouter(<LinkListItem label={EXAMPLE_TITLE} link={EXAMPLE_LINK} count={1} />)
-
-        const badge = screen.getByTestId('count-badge')
-
-        expect(badge).toBeInTheDocument()
     })
 
     it('handles a custom click event when passed through', async () => {